lil bits refactoring 2
