+++ MiracleGun/Invariants/GunShape.cs
