+++ MiracleGun/MiracleGun.csproj
@@ -47,13 +47,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="FortranPort\FortranFuncs.cs" />
-    <Compile Include="FortranPort\GasBound.cs" />
-    <Compile Include="FortranPort\GasNode.cs" />
+    <Compile Include="IdealGas\FortranFuncs.cs" />
+    <Compile Include="IdealGas\GasBound.cs" />
+    <Compile Include="IdealGas\GasLayer.cs" />
+    <Compile Include="IdealGas\GasCell.cs" />
     <Compile Include="Invariants\AbstractShape.cs" />
     <Compile Include="OvBallistic\OvBound.cs" />
     <Compile Include="OvBallistic\OvLayer.cs" />
-    <Compile Include="Invariants\MuzzleShape.cs" />
+    <Compile Include="Invariants\GunShape.cs" />
     <Compile Include="Invariants\GunPowder.cs" />
     <Compile Include="Invariants\GunPowderFactory.cs" />
     <Compile Include="OvBallistic\OvCell.cs" />
